# internal/adminapi/docs/swagger.json
{
    "components": {
        "schemas": {
            "adminapi.ConfigResponse": {
                "properties": {
                    "defaults": {
                        "$ref": "#/components/schemas/policy.Defaults"
                    },
                    "facts": {
                        "items": {
                            "$ref": "#/components/schemas/facts.Spec"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "groups": {
                        "items": {
                            "$ref": "#/components/schemas/adminapi.GroupViewElement"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "logging": {
                        "$ref": "#/components/schemas/policy.Logging"
                    }
                },
                "type": "object"
            },
...
            "adminapi.ErrorResponse": {
                "properties": {
                    "error": {
                        "description": "Error is a short, human-readable explanation.",
                        "example": "rebuild failed: group \"x\": rules must not be empty",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "adminapi.ItemListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/adminapi.ItemResponse"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "adminapi.RegisterResponse": {
                "properties": {
                    "payload": {
                        "type": "object"
                    },
                    "section": {
                        "example": "defaults",
                        "type": "string"
                    },
                    "stamp": {
                        "$ref": "#/components/schemas/crdt.Stamp"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Send 'Bearer \u003ctoken\u003e'. The token is the contents of --admin-token-file on the server.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "CRUD over the CRDT-backed sections of the policy: groups, facts, defaults, logging. The YAML loaded at boot is the floor; admin writes overlay it per-key and are gossiped to peer replicas. Every write rebuilds the effective config and either applies it atomically or quarantines the offending entry.",
        "title": "request-validator admin API",
        "version": "v1"
    },
...
    "paths": {
        "/api/v1/config": {
            "get": {
                "description": "Returns the policy the engine would evaluate right now (YAML floor + CRDT overrides), with sources annotated per group.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ConfigResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "merge or compile failed"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Effective compiled policy",
                "tags": [
                    "config"
                ]
            }
        },
        "/api/v1/defaults": {
            "delete": {
                "description": "After this call the YAML defaults are the effective ones again.",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Clear the defaults overlay",
                "tags": [
                    "defaults"
                ]
            },
            "get": {
                "description": "Returns 404 when no admin override is set; in that case the YAML defaults are in effect.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.RegisterResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "not set via admin api"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get the API-managed defaults overlay",
                "tags": [
                    "defaults"
                ]
            },
            "put": {
                "description": "Per-field overlay over the YAML defaults. Any field omitted from the body keeps its YAML value.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/adminapi.DefaultsRequest",
                                        "summary": "body",
                                        "description": "fields to override"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "fields to override",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.RegisterResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Override the defaults block",
                "tags": [
                    "defaults"
                ]
            }
        },
        "/api/v1/facts": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ItemListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List CRDT-managed facts",
                "tags": [
                    "facts"
                ]
            }
        },
...
        "/api/v1/groups": {
            "get": {
                "description": "Returns every group currently held in the CRDT store on this node. YAML-only groups are NOT included; use GET /api/v1/config for the merged effective view.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ItemListResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "missing or invalid bearer token"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List CRDT-managed groups",
                "tags": [
                    "groups"
                ]
            }
        },
...
        "/api/v1/logging": {
            "delete": {
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Clear the logging overlay",
                "tags": [
                    "logging"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.RegisterResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "not set via admin api"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get the API-managed logging overlay",
                "tags": [
                    "logging"
                ]
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/adminapi.LoggingRequest",
                                        "summary": "body",
                                        "description": "fields to override"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "fields to override",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.RegisterResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Override the logging block",
                "tags": [
                    "logging"
                ]
            }
        },
        "/api/v1/openapi.json": {
            "get": {
                "description": "Returns the generated openapi.json. Useful for client codegen.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminapi.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "OpenAPI 3.1 specification of this admin API",
                "tags": [
                    "meta"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ]
}